requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.2
torchaudio>=2.1.2
transformers>=4.38.0
datasets>=2.16.0
PyPDF2>=3.0.1
wandb>=0.16.0
tqdm>=4.66.0
numpy>=1.24.0
aiohttp>=3.8.0
accelerate>=0.25.0
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0 